+++ breeze-core/src/jmh/kotlin/core/StringSplitAndReplaceBenchmark.kt
@@ -12,19 +12,19 @@ import java.util.concurrent.*
 @Warmup(iterations = 5, time = 1)
 @Measurement(iterations = 5, time = 1)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
-class StringBenchmark {
+class StringSplitAndReplaceBenchmark {
 	//测试java和kotlin的String的split和replace方法的性能差异
 	//注意java的String的split和replace方法方法基本都是基于正则的，而kotlin的不是
 
-	//Benchmark                       Mode  Cnt    Score   Error  Units
-	//StringBenchmark.splitString1    avgt    5   58.501 ± 0.824  ns/op (kotlin) string.split(char)
-	//StringBenchmark.splitString2    avgt    5   71.052 ± 0.602  ns/op (kotlin) string.split(string)
-	//StringBenchmark.splitString3    avgt    5   72.372 ± 1.200  ns/op (java) string.split(regex)
-	//StringBenchmark.replaceString1  avgt    5   36.672 ± 0.858  ns/op (kotlin) string.replace(string,string)
-	//StringBenchmark.replaceString2  avgt    5    8.486 ± 0.120  ns/op (kotlin) string.replace(char,char)
-	//StringBenchmark.replaceString3  avgt    5    8.469 ± 0.159  ns/op (java) string.replace(char,char)
-	//StringBenchmark.replaceString4  avgt    5   37.419 ± 0.869  ns/op (java) string.replace(string,string)
-	//StringBenchmark.replaceString5  avgt    5  237.925 ± 3.552  ns/op (java) string.replaceAll(string,string)
+	//Benchmark                                     Mode  Cnt    Score   Error  Units
+	//StringSplitAndReplaceBenchmark.splitString1   avgt    5   58.501 ± 0.824  ns/op (kotlin) string.split(char)
+	//StringSplitAndReplaceBenchmark.splitString2   avgt    5   71.052 ± 0.602  ns/op (kotlin) string.split(string)
+	//StringSplitAndReplaceBenchmark.splitString3   avgt    5   72.372 ± 1.200  ns/op (java) string.split(regex)
+	//StringSplitAndReplaceBenchmark.replaceString1 avgt    5   36.672 ± 0.858  ns/op (kotlin) string.replace(string,string)
+	//StringSplitAndReplaceBenchmark.replaceString2 avgt    5    8.486 ± 0.120  ns/op (kotlin) string.replace(char,char)
+	//StringSplitAndReplaceBenchmark.replaceString3 avgt    5    8.469 ± 0.159  ns/op (java) string.replace(char,char)
+	//StringSplitAndReplaceBenchmark.replaceString4 avgt    5   37.419 ± 0.869  ns/op (java) string.replace(string,string)
+	//StringSplitAndReplaceBenchmark.replaceString5 avgt    5  237.925 ± 3.552  ns/op (java) string.replaceAll(string,string)
 
 	@Benchmark fun splitString1() = "1,2,3,4".split(',')
 	@Benchmark fun splitString2() = "1,2,3,4".split(",")