+++ breeze-core/src/main/kotlin/core/extension/UnstableExtensions.kt
@@ -9,28 +9,6 @@ import icu.windea.breezeframework.core.annotation.*
 import java.util.concurrent.*
 
 /**
- * 得到指定的一组值中第一个不为null的值，或者抛出异常。
- */
-@UnstableApi
-fun <T> coalesce(vararg values: T?): T {
-	for(value in values) {
-		if(value != null) return value
-	}
-	throw IllegalArgumentException("No non-null value in values.")
-}
-
-/**
- * 得到指定的一组值中第一个不为null的值，或者返回null。
- */
-@UnstableApi
-fun <T> coalesceOrNull(vararg values: T?): T? {
-	for(value in values) {
-		if(value != null) return value
-	}
-	return null
-}
-
-/**
  * 判断指定的关键字是否模糊匹配当前字符串。
  *
  * * 指定的关键字中的字符是否被当前字符串按顺序全部包含。